Fix type definition
