+++ types/index.d.ts
@@ -1,5 +1,4 @@
 export * from "./base";
 export * from "./client";
-export * from "./iteration";
 export * from "./person";
 export * from "./project";
